+++ examples/find_capabilities.py
@@ -11,8 +11,12 @@ lan_target = collector.configuration['target_register'].resolve("lan")
 async def doit():
     async for pkt, _, _ in lan_target.script(DeviceMessages.GetVersion()).run_with(FoundSerials()):
         if pkt | DeviceMessages.StateVersion:
-            cap = capability_for_ids(pkt.product, pkt.vendor)
-            print("{}: {}".format(pkt.serial, cap))
+            try:
+                cap = capability_for_ids(pkt.product, pkt.vendor)
+            except:
+                pass
+            else:
+                print("{}: {}".format(pkt.serial, cap))
 
 loop = collector.configuration["photons_app"].loop
 loop.run_until_complete(doit())